requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
python-multipart==0.0.20
python-magic==0.4.27; sys_platform != 'win32'
python-magic-bin==0.4.14; sys_platform == 'win32'
pikepdf==9.7.0
requests==2.32.3
mega.py==1.0.8
tenacity==9.1.2
python-dotenv==1.1.0
